Certificatet Data Extraction/Certificatet Data Extraction/requirements.txt
azure-ai-formrecognizer
azure-core
python-dotenv
google-generativeai
pandas
